--- mealie/routes/recipe/image_and_assets.py
@@ -1,69 +1,0 @@
-from shutil import copyfileobj
-
-from fastapi import Depends, File, Form, HTTPException, status
-from fastapi.datastructures import UploadFile
-from pydantic import BaseModel
-from slugify import slugify
-from sqlalchemy.orm.session import Session
-
-from mealie.db.db_setup import generate_session
-from mealie.repos.all_repositories import get_repositories
-from mealie.routes._base.routers import UserAPIRouter
-from mealie.schema.recipe import CreateRecipeByUrl, Recipe, RecipeAsset
-from mealie.services.recipe.recipe_data_service import RecipeDataService
-
-router = UserAPIRouter()
-
-
-class UpdateImageResponse(BaseModel):
-    image: str
-
-
-@router.post("/{slug}/image")
-def scrape_image_url(slug: str, url: CreateRecipeByUrl):
-    """Removes an existing image and replaces it with the incoming file."""
-
-    scrape_image(url.url, slug)
-
-
-@router.put("/{slug}/image", response_model=UpdateImageResponse)
-def update_recipe_image(
-    slug: str,
-    image: bytes = File(...),
-    extension: str = Form(...),
-    session: Session = Depends(generate_session),
-):
-    """Removes an existing image and replaces it with the incoming file."""
-    db = get_repositories(session)
-    write_image(slug, image, extension)
-    new_version = db.recipes.update_image(slug, extension)
-
-    return UpdateImageResponse(image=new_version)
-
-
-@router.post("/{slug}/assets", response_model=RecipeAsset)
-def upload_recipe_asset(
-    slug: str,
-    name: str = Form(...),
-    icon: str = Form(...),
-    extension: str = Form(...),
-    file: UploadFile = File(...),
-    session: Session = Depends(generate_session),
-):
-    """Upload a file to store as a recipe asset"""
-    file_name = slugify(name) + "." + extension
-    asset_in = RecipeAsset(name=name, icon=icon, file_name=file_name)
-    dest = Recipe(slug=slug).asset_dir.joinpath(file_name)
-
-    with dest.open("wb") as buffer:
-        copyfileobj(file.file, buffer)
-
-    if not dest.is_file():
-        raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    db = get_repositories(session)
-
-    recipe: Recipe = db.recipes.get(slug)
-    recipe.assets.append(asset_in)
-    db.recipes.update(slug, recipe.dict())
-    return asset_in
